==> EndToEndProjects/FraudDetection_AutoInsurance/01_DataPrep_Wrangler_FeatureStore.ipynb <==
# Prepare Data in this Step
 - ## Data Wrangler: To Make transformation and process the data
 - ## Sagemaker Feature Store: To store the features
   
1. Use Data Wrangler UI to transform and preprocess the data, create .flow files.
2. Use .flow files to create the processed datasets.
3. Store the features (processed dataset) in feature store
4. FeatureGroup Creation, Ingest batch data and retrive records

References:
 - https://github.com/aws/amazon-sagemaker-examples/blob/4534bff4b5b5062af5789d98c4ddca01b0cb5d1f/end_to_end/fraud_detection/0-AutoClaimFraudDetection.ipynb
 - https://docs.aws.amazon.com/sagemaker/latest/dg/feature-store-introduction-notebook.html
 - https://boto3.amazonaws.com/v1/documentation/api/1.20.9/reference/services/sagemaker-featurestore-runtime.html



from sagemaker import get_execution_role
from sagemaker.session import Session
import sagemaker
import boto3
import pandas as pd
import numpy as np 
from time import strftime, gmtime
import time
from sagemaker.feature_store.feature_group import FeatureGroup

sagemaker_session = Session()
REGION = sagemaker_session.boto_region_name
BUCKET = sagemaker_session.default_bucket()
PREFIX = "FraudDetection_AutoInsurance"

s3_client = boto3.client("s3")
sagemaker_client = boto3.client("sagemaker")
iam_client = boto3.client("iam")
print(REGION)

"""
Note: if you are not running this notebook from SageMaker Studio or SageMaker Classic Notebooks you will need to instanatiate 
the sagemaker_execution_role_name with an AWS role that has SageMakerFullAccess and SageMakerFeatureStoreFullAccess
"""
try:
    ROLE = sagemaker.get_execution_role()
    print(ROLE)
except:
    print("Need to set role manually")
    exec_role_name = "AmazonSageMaker-ExecutionRole-20250401T145997"
    sagemaker_role = iam_client.get_role(RoleName=exec_role_name)["Role"]["Arn"]
    print(f"\n instantiating sagemaker_role with supplied role name : {sagemaker_role}")

account_id = boto3.client("sts").get_caller_identity()["Account"]
print(account_id)

# Upload Files to S3 Buckets
#s3_client.upload_file(Filename="Data/customers.csv", Bucket=BUCKET, Key=f"{PREFIX}/data/raw/customers.csv")
#s3_client.upload_file(Filename="Data/claims.csv", Bucket=BUCKET, Key=f"{PREFIX}/data/raw/claims.csv")

### USE DATA WRANGLER UI TO TRANSFORM THE ABOVE RAW DATA AND GENERATE THE BELOW PROCESSED DATA
**I am skipping for this now. But will learn more about this step

s3_uri = f"s3://{BUCKET}"

claims_data = pd.read_csv(f"{s3_uri}/{PREFIX}/data/claims_preprocessed.csv")
customers_data = pd.read_csv(f"{s3_uri}/{PREFIX}/data/customers_preprocessed.csv")
print(claims_data.dtypes)
print(customers_data.dtypes)

print(claims_data.columns, "\n")
print(customers_data.columns, "\n")
print()
print("Policy id is the unique identifier in both cases")
print(claims_data.shape[0], claims_data['policy_id'].nunique())
print(customers_data.shape[0], customers_data['policy_id'].nunique())

### Create Two Feature Groups for claims and customers.
Feature Groups is equivalent to tables/datasets.

customers_feature_group_name = "customers-feature-group-" + strftime("%d-%H-%M-%S", gmtime())
claims_feature_group_name = "claims-feature-group-" + strftime("%d-%H-%M-%S", gmtime())
print(customers_feature_group_name, "\n", claims_feature_group_name)

# Initiate Feature Group objects. Each feature group reprsents a table.
customers_feature_group = FeatureGroup(customers_feature_group_name, sagemaker_session)
claims_feature_group = FeatureGroup(claims_feature_group_name, sagemaker_session)

# Add event time in the datasets, that is required
customers_data['EventTime'] = time.time()
claims_data['EventTime'] = time.time()
print(customers_data['EventTime'])

### Load Feature Definitions into the feature groups

customers_feature_group.load_feature_definitions(data_frame=customers_data)
claims_feature_group.load_feature_definitions(data_frame=claims_data)

### The following calls create to create two feature groups, customers_feature_group and orders_feature_group, respectively:
Point to note, only the necessary structure will be created on "Feature Store" we still need to load the data yet.


customers_feature_group.create(
    s3_uri = f"{s3_uri}/{PREFIX}/data/feature_store",
    record_identifier_name='policy_id',
    event_time_feature_name="EventTime",
    role_arn=ROLE,
    enable_online_store=True
)


claims_feature_group.create(
    s3_uri = f"{s3_uri}/{PREFIX}/data/feature_store",
    record_identifier_name='policy_id',
    event_time_feature_name="EventTime",
    role_arn=ROLE,
    enable_online_store=True
)

print(customers_feature_group.describe())

### List Feature Groups

print(sagemaker_session.boto_session.client('sagemaker').list_feature_groups())
print()
print(boto3.client('sagemaker').list_feature_groups())

def check_feature_group_status(feature_group):
    status = feature_group.describe().get("FeatureGroupStatus")
    while status == "Creating":
        print("Waiting for Feature Group to be Created")
        time.sleep(5)
        status = feature_group.describe().get("FeatureGroupStatus")
    print(f"FeatureGroup {feature_group.name} successfully created.")

check_feature_group_status(customers_feature_group)
check_feature_group_status(claims_feature_group)

### Upload data into feature groups

customers_feature_group.ingest(data_frame=customers_data, max_workers=3, wait=True) 
claims_feature_group.ingest(data_frame=claims_data, max_workers=3, wait=True) 

### If feature group already exists then initialize the feature group object and get the data

print(sagemaker_session.boto_session.client('sagemaker').list_feature_groups())
print()
feature_groups = boto3.client('sagemaker').list_feature_groups()

customers_feature_group_name = feature_groups['FeatureGroupSummaries'][0]['FeatureGroupName']
claims_feature_group_name = feature_groups['FeatureGroupSummaries'][1]['FeatureGroupName']

claims_feature_group = FeatureGroup(name=claims_feature_group_name, sagemaker_session=sagemaker_session)
customers_feature_group = FeatureGroup(name=customers_feature_group_name, sagemaker_session=sagemaker_session)
claims_feature_group.describe()

### Use Athena query to load the datasets

claims_query = claims_feature_group.athena_query()
print(claims_query.table_name, "\n", claims_query.database)
customers_query = customers_feature_group.athena_query()
print(customers_query.table_name, "\n", customers_query.database)

query1 = f"SELECT * FROM {customers_query.table_name}"
query2 = f"SELECT * FROM {claims_query.table_name}"

claims_query.run(query1, output_location=f"s3://{BUCKET}/FraudDetection_AutoInsurance/data/query_results")
claims_query.wait() 
customers_dataset = claims_query.as_dataframe()
claims_query.run(query2, output_location=f"s3://{BUCKET}/FraudDetection_AutoInsurance/data/query_results")
claims_query.wait() 
claims_dataset = claims_query.as_dataframe()

print(claims_dataset['is_deleted'].value_counts())
print(claims_dataset['policy_id'].nunique() == claims_data['policy_id'].shape[0])
print(customers_dataset['is_deleted'].value_counts())
print(customers_dataset['policy_id'].nunique() == claims_data['policy_id'].shape[0])

# Drop the logically deleted records, and keep the latest records only based on the metadata columns. Delete the metadata columns in last

keep_latest = lambda df: df.sort_values('write_time',ascending=False).drop_duplicates(['policy_id'], keep='first').reset_index(drop=True)
claims_dataset = keep_latest(claims_dataset[claims_dataset['is_deleted']==False])
customers_dataset = keep_latest(customers_dataset[customers_dataset['is_deleted']==False])

metadata_columns = ['write_time', 'api_invocation_time', 'is_deleted']
dataset = pd.merge(
    claims_dataset.drop(columns=metadata_columns),
    customers_dataset.drop(columns=metadata_columns),
    on=['policy_id']
)
print(claims_dataset.shape) 
print(customers_dataset.shape) 
print(dataset.shape)
print(dataset.dtypes)


dataset_col_order = ['fraud']  + list(dataset.drop(["fraud", "policy_id"], axis=1).columns)
train = dataset.sample(frac=.80, random_state=0)[dataset_col_order]
test = dataset.drop(train.index)[dataset_col_order]
train.to_csv("data/train.csv", index=False)
test.to_csv("data/test.csv", index=False)
dataset.to_csv("data/dataset.csv", index=True)

# Upload the data to S3
s3_client = sagemaker_session.boto_session.client("s3")
s3_client.upload_file('data/train.csv', Bucket=BUCKET, Key="FraudDetection_AutoInsurance/data/train.csv")
s3_client.upload_file('data/test.csv', Bucket=BUCKET, Key="FraudDetection_AutoInsurance/data/test.csv")
s3_client.upload_file('data/dataset.csv', Bucket=BUCKET, Key="FraudDetection_AutoInsurance/data/dataset.csv")


print(train.select_dtypes(['int','float']).shape)
print(train.shape)
