==> EndToEndProjects/FraudDetection_AutoInsurance/03_SagmakerClarify.ipynb <==
# Sagamaker Clarify

- **ReadtheDocs link:** https://sagemaker-examples.readthedocs.io/en/latest/sagemaker-clarify/index.html
- **Pre-Training Bias:** https://docs.aws.amazon.com/sagemaker/latest/dg/clarify-detect-data-bias.html

import sagemaker
import boto3

ROLE = sagemaker.get_execution_role()
print(ROLE)

boto_session = boto3.session.Session()
REGION = boto_session.region_name
print(REGION)

sagemaker_session = sagemaker.session.Session(boto_session=boto_session)
print(sagemaker_session.boto_region_name)
BUCKET = sagemaker_session.default_bucket()
print(BUCKET)

KEY = "FraudDetection_AutoInsurance"

sagemaker_client = boto_session.client("sagemaker")
s3_client = boto_session.client("s3")

### First check if there is a model registered 
### Create the model from the last training job that we had run

model_name = "sagemaker-xgboost-v3"
models = sagemaker_client.list_models(NameContains=model_name)['Models']
print(models)
if not models:
    
    training_job_name = 'sagemaker-xgboost-2025-06-11-13-18-27-251'  #'sagemaker-xgboost-2025-05-30-09-09-33-714'
    training_job_info = sagemaker_client.describe_training_job(TrainingJobName=training_job_name)
    print(training_job_info['TrainingJobName'])
    print(training_job_info["AlgorithmSpecification"]["TrainingImage"])
    
    model1 = sagemaker_session.create_model_from_job(
        name=model_name, 
        training_job_name=training_job_name, 
        role=ROLE, 
        image_uri=training_job_info["AlgorithmSpecification"]["TrainingImage"]
    )
    print("Creating Model")

#### Now create a sagmaker clarify process

clarify_instance_count = 1
clarify_instance_type = "ml.c4.xlarge"

train_data_uri = f"s3://{BUCKET}/{KEY}/data/train.csv"
test_data_uri = f"s3://{BUCKET}/{KEY}/data/test.csv"

import pandas as pd
train_data = pd.read_csv(train_data_uri)
print(train_data.columns)  
print(train_data['fraud'].value_counts())

bias_report_1_output_path = f"s3://{BUCKET}/{KEY}/clarify-bias"
print(bias_report_1_output_path)

## Create a clarify process, prividiing data config, model config, predictions config and the bias config

clarify_processor = sagemaker.clarify.SageMakerClarifyProcessor(
    role=ROLE, 
    instance_count=clarify_instance_count, 
    instance_type=clarify_instance_type, 
    sagemaker_session=sagemaker_session
)
print(clarify_processor)

train_data.dtypes

# Dataconfig - Basic Information about the input data i.e. input data path, output report path, target_label, columns name list, dataset type i.e. csv/json etc
bias_data_config = sagemaker.clarify.DataConfig(
    s3_data_input_path=train_data_uri, 
    s3_output_path=bias_report_1_output_path, 
    label='fraud',
    headers=train_data.columns.tolist(),
    dataset_type="text/csv"
)

# Model Config: Basic Information about the model i.e. model_name to assess bias for, instance_type to run model on, input and output of the mdoel. 
bias_model_config = sagemaker.clarify.ModelConfig(
    model_name=model_name,
    instance_count=clarify_instance_count,
    instance_type=clarify_instance_type,
    accept_type="text/csv"#,
    #content_type="text/csv"
)

# Model Predicted Lable Config i.e. define model score threshold
predictions_config = sagemaker.clarify.ModelPredictedLabelConfig(probability_threshold=0.02)

# Bias Config
bias_config = sagemaker.clarify.BiasConfig(
    label_values_or_threshold=[1],
    facet_name='customer_gender_female',
    facet_values_or_threshold=[1],
    group_name='customer_age'
)

clarify_processor.run_bias(
    data_config=bias_data_config,
    model_config=bias_model_config,
    bias_config=bias_config,
    model_predicted_label_config=predictions_config,
    pre_training_methods="all",
    post_training_methods="all"
)
