==> EndToEndProjects/FraudDetection_AutoInsurance/04_ModelRegistry_And_Deploy.ipynb <==

# 1. Deposit Model and Lineage in SageMaker Model Registry
# 2. Deploy to an endpoint

Once a useful model has been trained and its artifacts properly associated, the next step is to save the model in a registry for future reference and possible deployment.

References: 
- https://docs.aws.amazon.com/sagemaker/latest/dg/model-registry.html
- https://sagemaker-examples.readthedocs.io/en/latest/end_to_end/fraud_detection/2-lineage-train-assess-bias-tune-registry-e2e.html

#### What Is Model Registry in SageMaker?
SageMaker Model Registry is a centralized repository to:

- Catalog models for production.
- Manage model versions.
- Associate metadata, such as training metrics, with a model.
- View information from Amazon SageMaker Model Cards in your registered models.
- View model lineage for traceability and reproducibility.
- Define a staging construct that models can progress through for your model lifecycle.
- Manage the approval status of a model.
- Deploy models to production.
- Automate model deployment with CI/CD.

Share models with other users.
#### But When Should You Register a Model?

📌 1. Versioning

- Model Registry keeps track of multiple versions.
- You can roll back to a previous version easily.
- Each version stores metadata like creation time, metrics, and approval status.

📌 2. Approval Workflow
- You can set a model’s status as: PendingManualApproval (Default), Approved, Rejected.
- Good for human review, audit, or CI/CD pipelines.
- Prevents accidental deployment of unverified models.

📌 3. Audit & Traceability
- Logs who registered the model, which training job it came from, and what data was used.
- Helps in regulated environments like finance, healthcare, or defense.

📌 4. Production-readiness
- Enables stage promotion (e.g., dev → staging → production).
- Integrates with SageMaker Pipelines for full MLOps automation.

📌 5. Team collaboration
- Shared model group across teams lets everyone see:
- Latest approved model
- Metrics, parameters, metadata
- History of changes

References:
 - https://sagemaker-examples.readthedocs.io/en/latest/end_to_end/fraud_detection/2-lineage-train-assess-bias-tune-registry-e2e.html

import sagemaker
import boto3

REGION = sagemaker.Session().boto_region_name
boto_session = boto3.Session(region_name=REGION)
s3_client = boto_session.client('s3')
sagemaker_client = boto3.client('sagemaker', region_name=REGION)

sagemaker_session = sagemaker.session.Session(boto_session=boto_session, sagemaker_client=sagemaker_client)

BUCKET = sagemaker_session.default_bucket()
ROLE = sagemaker.get_execution_role()
PREFIX = "FraudDetection_AutoInsurance"
print(REGION)
print(BUCKET)
print(ROLE)

### If no model is found with the name, then create it from the training job


model_name = "sagemaker-xgboost-v3"
training_job_name = 'sagemaker-xgboost-2025-06-11-13-18-27-251'

training_job_info = sagemaker_client.describe_training_job(TrainingJobName=training_job_name)
print(training_job_info['TrainingJobName'])
print(training_job_info["AlgorithmSpecification"]["TrainingImage"])
training_job_info["ModelArtifacts"]["S3ModelArtifacts"]    

models = sagemaker_client.list_models(NameContains=model_name)['Models']
print(models)
if not models:
    # If no model is found with the name, then create it from the training job
    
    model1 = sagemaker_session.create_model_from_job(
        name=model_name, 
        training_job_name=training_job_name, 
        role=ROLE, 
        image_uri=training_job_info["AlgorithmSpecification"]["TrainingImage"]
    )
    print("Creating Model")

## Now Registered the model in the Model Registry

### Create Model Package Group
A Model Package Groups holds multiple versions or iterations of a model. Though it is not required to create them for every model in the registry, they help organize various models which all have the same purpose and provide automatic versioning.

The following code will create a model package group i.e. "SageMaker Studio > Models > Registered Models > FraudDetection AutoInsurance"


model_package_group_name = PREFIX.replace('_','-')
#sagemaker_client.list_model_package_groups(NameContains=model_package_group_name)
matching_group = sagemaker_client.list_model_package_groups(NameContains=model_package_group_name)['ModelPackageGroupSummaryList']
if matching_group:
    print(f'Using existing Model Package Group: {model_package_group_name}')
else:
    mpg_input_dict = {'ModelPackageGroupName': model_package_group_name, 'ModelPackageGroupDescription': 'Auto Insurance E2E'}
    mpg_response = sagemaker_client.create_model_package_group(**mpg_input_dict)
    print(f'Create Model Package Group {model_package_group_name}: SUCCESSFUL')

### Create Model Package

#### Define the model inference specification
- Containers: i.e. container image, s3 location of model output i.e. model.tar.gz
- Request Type: What format the endpoint will accept the input for inference
- Response Type: Response format.

**Point to note:** This settings serve a declarative/documentation role, not an enforcement one at runtime for SageMaker’s Predictor client. So you need to define serializers and deserializers at the time of predictions

model_package_inference_specification = {
    "InferenceSpecification": {
        "Containers": [{
        "Image": training_job_info["AlgorithmSpecification"]["TrainingImage"],
        "ModelDataUrl": training_job_info["ModelArtifacts"]["S3ModelArtifacts"]
        }],
        "SupportedContentTypes": ["text/csv"],
        "SupportedResponseMIMETypes": ["text/csv"],
    }
}
print(model_package_inference_specification)


create_model_package_input_dict = {
    "ModelPackageGroupName": model_package_group_name,
    "ModelPackageDescription": "Model to detect fraud in auto-insurance",
    "ModelApprovalStatus": "PendingManualApproval" # Defaul value
}
create_model_package_input_dict.update(model_package_inference_specification)

create_model_package_response = sagemaker_client.create_model_package(**create_model_package_input_dict)
model_package_arn = create_model_package_response["ModelPackageArn"]
print('ModelPackage Version ARN : {}'.format(model_package_arn))

#### List and Describe Model Packages

print(sagemaker_client.list_model_package_groups())
print()
print(sagemaker_client.list_model_packages(ModelPackageGroupName='FraudDetection-AutoInsurance'))
print()
print(sagemaker_client.describe_model_package(ModelPackageName='arn:aws:sagemaker:us-east-1:205930620783:model-package/FraudDetection-AutoInsurance/1'))
print()
#print(sagemaker_client.describe_model_package(ModelPackageName='arn:aws:sagemaker:us-east-1:205930620783:model-package/FraudDetection-AutoInsurance/1')['ModelPackageName'])



### Model Registry Collections
You can use Collections to group registered models that are related to each other and organize them in hierarchies to improve model discoverability at scale. With Collections, you can organize registered models that are associated with one another. For example, you could categorize your models based on the domain of the problem they solve as Collections titled NLP-models, CV-models, or Speech-recognition-models. To organize your registered models in a tree structure, you can nest Collections within each other.

**For now we are skipping it**

## Model Deployment to an endpoint for real-time inference

## First Approve the model to deploy
update_model_package_dict = {
    'ModelPackageArn': 'arn:aws:sagemaker:us-east-1:205930620783:model-package/FraudDetection-AutoInsurance/1',
    'ModelApprovalStatus': 'Approved'
}
sagemaker_client.update_model_package(**update_model_package_dict)

endpoint_name = f"{model_package_group_name}-endpoint"
print(endpoint_name)
endpoint_instance_count = 1
endpoint_instance_type = "ml.t2.medium"#"ml.m4.xlarge"


from sagemaker import ModelPackage

model_package = ModelPackage(
    role=ROLE, 
    model_package_arn='arn:aws:sagemaker:us-east-1:205930620783:model-package/FraudDetection-AutoInsurance/1',
    sagemaker_session=sagemaker_session
)

predictor = model_package.deploy(
    endpoint_name=endpoint_name,
    instance_type=endpoint_instance_type,
    initial_instance_count=1
)

help(model_package.deploy)
